test(content): colocate all source tests
Move the remaining content loader, artifact, and path tests beside their implementations. Keep fixture-only dependencies inside import.meta.vitest blocks and explicitly list safe source modules so the test runner does not execute build scripts with top-level side effects.
