--- packages/content/scripts/content-paths.test.ts
@@ -1,8 +1,0 @@
-import path from 'node:path';
-import { blogDirectory } from '../src/paths.ts';
-
-describe('blogDirectory', () => {
-	it('targets the content workspace', () => {
-		expect(blogDirectory('/workspace')).toBe(path.join('/workspace', 'src', 'blog'));
-	});
-});
